// api/model/chargePointProtocol.ts
/**
 * CPMS API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ChargePointProtocol = 'Undefined' | 'OCPP_1_2' | 'OCPP_1_5' | 'OCPP_1_6_J' | 'OCPP_1_6_S' | 'OCPP_2_0';

export const ChargePointProtocol = {
    Undefined: 'Undefined' as ChargePointProtocol,
    OCPP_1_2: 'OCPP_1_2' as ChargePointProtocol,
    OCPP_1_5: 'OCPP_1_5' as ChargePointProtocol,
    OCPP_1_6_J: 'OCPP_1_6_J' as ChargePointProtocol,
    OCPP_1_6_S: 'OCPP_1_6_S' as ChargePointProtocol,
    OCPP_2_0: 'OCPP_2_0' as ChargePointProtocol
};