// api/model/chargePointStatus.ts
/**
 * AdminApp API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ChargePointStatus = 'Unavailable' | 'Available' | 'Occupied' | 'Faulted' | 'Unknown';

export const ChargePointStatus = {
    Unavailable: 'Unavailable' as ChargePointStatus,
    Available: 'Available' as ChargePointStatus,
    Occupied: 'Occupied' as ChargePointStatus,
    Faulted: 'Faulted' as ChargePointStatus,
    Unknown: 'Unknown' as ChargePointStatus
};