// api/model/chargeSessionStatus.ts
/**
 * CPMS API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ChargeSessionStatus = 'Undefined' | 'Pending' | 'Requested' | 'Accepted' | 'Charging' | 'Completed' | 'Faulted' | 'Rejected' | 'Cancelled';

export const ChargeSessionStatus = {
    Undefined: 'Undefined' as ChargeSessionStatus,
    Pending: 'Pending' as ChargeSessionStatus,
    Requested: 'Requested' as ChargeSessionStatus,
    Accepted: 'Accepted' as ChargeSessionStatus,
    Charging: 'Charging' as ChargeSessionStatus,
    Completed: 'Completed' as ChargeSessionStatus,
    Faulted: 'Faulted' as ChargeSessionStatus,
    Rejected: 'Rejected' as ChargeSessionStatus,
    Cancelled: 'Cancelled' as ChargeSessionStatus
};