// api/model/chargingLevel.ts
/**
 * CPMS API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ChargingLevel = 'Undefined' | 'Level1' | 'Level2' | 'Level3';

export const ChargingLevel = {
    Undefined: 'Undefined' as ChargingLevel,
    Level1: 'Level1' as ChargingLevel,
    Level2: 'Level2' as ChargingLevel,
    Level3: 'Level3' as ChargingLevel
};